remember Me,Logout and modals

+++ Angular/scripts/controllers/userDashboardController.js
@@ -2,10 +2,11 @@
 
     var app = angular.module("school");
 
-    app.controller("UserDashboardController", function ($scope, $rootScope, DataService, schConfig) {
+    app.controller("UserDashboardController", function ($scope, $rootScope, DataService, schConfig, $modal) {
         var dataSet = "userreservations";
         $scope.repeatingTypes = schConfig.repeatingType;
         fetchData();
+       
 
         function fetchData() {
             DataService.list(dataSet, function (data) {
@@ -13,15 +14,36 @@
             });
         }
 
+
         $scope.cancelReservation = function (item) {
+
             var index = $scope.dashboard.activeReservations.indexOf(item);
             $scope.eventId = item.id;
+            console.log($scope.eventId);
             getEvent($scope.eventId);
-            if ($scope.reservationEvent != undefined) {
-                DataService.remove("events", $scope.reservationEvent.id, function (data) { });
-                $scope.dashboard.activeReservations.splice(index, 1);
+            $scope.confirmed = {
+                isConfirmed:false
             }
-            $scope.$apply();
+
+            var modalInstance = $modal.open({
+                templateUrl: 'views/modals/cancelResModal.html',
+                controller: 'CancelResCtrl',
+                windowClass: 'app-modal-window',
+                backdrop: 'static',
+                size: 'sm',
+                resolve : {
+                   confirmed : function() {
+                        return $scope.confirmed;
+                    }
+                }
+            }).result.then(function(result) {
+                $scope.isConfirmed = result;
+                if ($scope.reservationEvent != undefined && $scope.isConfirmed) {
+                    DataService.remove("events", $scope.reservationEvent.id, function (data) { });
+                    $scope.dashboard.activeReservations.splice(index, 1);
+                }
+            });
+         
         };
 
         $scope.extendReservation = function (item) {
@@ -32,22 +54,23 @@
                 repeatingType: $scope.repeatingTypes.indexOf(0),
                 frequency: 0
             }
-            $scope.modalShow = true;
+            var modalInstance = $modal.open({
+                templateUrl: 'views/modals/extendModal.html',
+                controller:'ExtendModalCtrl',
+                windowClass: 'app-modal-window',
+                backdrop: 'static',
+                size: 'md',
+                scope: $scope
+            });
         };
-
-        $scope.saveData = function () {
-            DataService.create("eventextends", $scope.eventExtend, function (data) { });
-            $scope.modalShow = false;
-        };
-
-        $scope.cancel = function() {
-            $scope.modalShow = false;
-        };
-
+  
+     
         function getEvent(id) {
-            DataService.read("events", id , function (data) {
+            DataService.read("events", id, function (data) {
                 $scope.reservationEvent = data;
             });
         };
     });
+
+   
 }());
